Update data.go
Change in code to use the new setting and allow to have a bigger than 480 channel list if not using a Plex Client, is set to max 999 for performance considerations. I didn't modified the settings UI for this new setting, it can be done (I don't know how) but still it'll require to restart the server to apply it.